sandbox: extract shared BackendProtocol conformance suite
